refactoring/#15-shad-cn-integration разложил по каталогам
