+++ src/components/settings/UploadBackup.tsx
@@ -1,5 +1,5 @@
 import { HardDriveUpload } from 'lucide-react';
-import { persistor } from '../model/store.ts';
+import { persistor } from '../../model/store.ts';
 import { Button } from '@/components/ui/button.tsx';
 
 export const UploadBackup = () => {
